backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# AI/ML dependencies
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
transformers==4.35.2

# API clients for external services
groq==0.4.1
tavily-python==0.3.1

# Image processing
opencv-python==4.8.1.78
imageio==2.31.6
matplotlib==3.8.2
seaborn==0.13.0

# PDF and Report Generation
reportlab==4.0.7
fpdf2==2.7.6
jinja2==3.1.2

# Multi-language Support
langdetect==1.0.9
googletrans==4.0.0rc1

# Data handling
pandas==2.1.4
requests==2.31.0

# System monitoring
psutil==5.9.6

# Environment and configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
flake8==6.1.0

# Logging and monitoring
structlog==23.2.0

# Database (optional)
sqlalchemy==2.0.23
alembic==1.13.1

# Caching (optional)
redis==5.0.1

# File handling
aiofiles==23.2.1

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3

# DICOM support (optional)
pydicom==2.4.3
